pyproject.toml
[project]
name = "m365-update-manager"
version = "0.1.0"
description = "Lagged deployment manager for Microsoft 365 macOS updates"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "David Crosby", email = "bing@bing-bong.co.uk" }
]

dependencies = [
    "requests>=2.31.0",
    "azure-storage-blob>=12.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "types-requests>=2.31.0",
    "pre-commit>=3.6.0",
]

[project.scripts]
check-updates = "check_updates:main"
promote = "promote:main"

[project.urls]
Repository = "https://github.com/david-crosby/m365-update-manager"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term-missing"
pythonpath = ["."]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
check_untyped_defs = true
no_implicit_optional = true
